requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
boto3==1.23.9
botocore==1.26.9
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
Django==3.2.12
django-cors-headers==3.11.0
django-filter==21.1
django-sslserver==0.22
django-storages==1.12.3
django-utils==0.0.2
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
h11==0.13.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jmespath==1.0.0
outcome==1.1.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.5.2
selenium==4.1.3
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.20.0
trio-websocket==0.9.2
Twisted==22.2.0
txaio==22.2.1
typing_extensions==4.1.1
urllib3==1.26.9
wsproto==1.1.0
zope.interface==5.4.0
